deps/v8/test/inspector/debugger/return-break-locations-expected.txt

Return break locations within function

Running test: testTailCall
[
    [0] : {
        columnNumber : 0
        lineNumber : 6
        scriptId : <scriptId>
        type : return
    }
]
